// src/api-generated/model/aprovar-proposta-dto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * API Horários Acadêmicos
 * Documentação da API para o Sistema de Elaboração de Horário e Atribuição de Disciplinas
 * OpenAPI spec version: 1.0
 */

export interface AprovarPropostaDto {
  /** Observações opcionais do diretor sobre a aprovação */
  observacoesDiretor?: string
}